python scripts: BitBoxBase python module.
Extracted the shared API calls between BB02 and BBBase into a new module
(communication.bitbox_api_protocol). Created a new "bitboxbase" module
which implements these shared APIs (e.g. reboot), together with
BitBoxBase-specific APIs.
